+++ Assets/Scripts/architecture_business/use_cases/enemy_spacecraft/life/SpacecraftEnemyCurrentLifeUseCase.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public interface SpacecraftEnemyCurrentLifeUseCase {
-    int invoke(IdentificatorModel identificator);
+    float invoke(IdentificatorModel identificator);
 }
 
 public class SpacecraftEnemyCurrentLifeUseCaseImpl : SpacecraftEnemyCurrentLifeUseCase
 {
     private SpacecraftEnemyLifeRepository repo = new SpacecraftEnemyLifeRepositoryImpl();
 
-    public int invoke(IdentificatorModel identificator) => repo.currentLife(identificator);
+    public float invoke(IdentificatorModel identificator) => repo.currentLife(identificator);
 }
